Move util methods to class Columns
