+++ src/main/java/marquez/db/mappers/DatasetRowMapper.java
@@ -1,8 +1,5 @@
 package marquez.db.mappers;
 
-import static marquez.db.Utils.toInstantOrNull;
-import static marquez.db.Utils.toUuidOrNull;
-
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.UUID;
@@ -19,12 +16,12 @@ public final class DatasetRowMapper implements RowMapper<DatasetRow> {
     return DatasetRow.builder()
         .uuid(results.getObject(Columns.ROW_UUID, UUID.class))
         .createdAt(results.getTimestamp(Columns.CREATED_AT).toInstant())
-        .updatedAt(toInstantOrNull(results.getTimestamp(Columns.UPDATED_AT)))
+        .updatedAt(Columns.toInstantOrNull(results.getTimestamp(Columns.UPDATED_AT)))
         .namespaceUuid(results.getObject(Columns.NAMESPACE_UUID, UUID.class))
         .dataSourceUuid(results.getObject(Columns.DATA_SOURCE_UUID, UUID.class))
         .urn(results.getString(Columns.URN))
         .description(results.getString(Columns.DESCRIPTION))
-        .currentVersion(toUuidOrNull(results.getString(Columns.CURRENT_VERSION_UUID)))
+        .currentVersion(Columns.toUuidOrNull(results.getString(Columns.CURRENT_VERSION_UUID)))
         .build();
   }
 }
