+++ src/main/java/marquez/db/mappers/JobRowMapper.java
@@ -1,7 +1,5 @@
 package marquez.db.mappers;
 
-import static marquez.db.Utils.toList;
-
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.UUID;
@@ -21,8 +19,8 @@ public final class JobRowMapper implements RowMapper<Job> {
         results.getString(Columns.LOCATION),
         results.getObject(Columns.NAMESPACE_UUID, UUID.class),
         results.getString(Columns.DESCRIPTION),
-        toList(results.getArray(Columns.INPUT_DATASET_URNS)),
-        toList(results.getArray(Columns.OUTPUT_DATASET_URNS)),
+        Columns.toList(results.getArray(Columns.INPUT_DATASET_URNS)),
+        Columns.toList(results.getArray(Columns.OUTPUT_DATASET_URNS)),
         results.getTimestamp(Columns.CREATED_AT));
   }
 }
