+++ src/main/java/marquez/db/mappers/JobVersionRowMapper.java
@@ -1,7 +1,5 @@
 package marquez.db.mappers;
 
-import static marquez.db.Utils.toUuidOrNull;
-
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.UUID;
@@ -20,7 +18,7 @@ public final class JobVersionRowMapper implements RowMapper<JobVersion> {
         results.getObject(Columns.JOB_UUID, UUID.class),
         results.getString(Columns.LOCATION),
         results.getObject(Columns.VERSION, UUID.class),
-        toUuidOrNull(results.getString(Columns.LATEST_JOB_RUN_UUID)),
+        Columns.toUuidOrNull(results.getString(Columns.LATEST_JOB_RUN_UUID)),
         results.getTimestamp(Columns.CREATED_AT),
         results.getTimestamp(Columns.UPDATED_AT));
   }
